images/exporter/pyproject.toml
[project]
name = "exporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "mkdocs==1.6.1",
    "mkdocs-katex-ssr>=1.0.9",
    "mkdocs-material",
    "mkdocs-minify-html-plugin>=0.3.9",
    "mkdocs-table-reader-plugin>=3.1.0",
    "pymdown-extensions>=10.21",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "toml>=0.10.2",
]

[tool.uv.sources]
mkdocs-material = { git = "https://github.com/raineblog/mkdocs-material.git" }
